requirements.txt
fastapi[standard]
attrs
pydantic
uvicorn
sqlmodel
sqlalchemy
# async
aiosqlite
greenlet
# auth
python-multipart
pyjwt
passlib[bcrypt]
# testing
pytest
